transcription_service_controller/pyproject.toml
[tool.poetry]
name = "transcription-service-controller"
version = "0.1.0"
description = "Transcription Service Controller API"
authors = ["usnar"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.101.1"
uvicorn = {extras = ["standard"], version = "0.22.0"}
celery = "5.3.0"
redis = "4.6.0"
aiohttp = "3.12.13"
python-dotenv = "1.0.0"
psycopg-pool = "3.2.6"
requests = "2.31.0"
psycopg = {extras = ["binary"], version = "*"}
sqlalchemy = "2.0.20"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-asyncio = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"